backend/requirements.txt
django
djangorestframework
mysqlclient
Pillow
django-cors-headers
python-decouple